Verify-Up
This program tests whether a site is up or not by downloading a known
file from the remote server. If the file is able to be downloaded, the
site is considered to be up, and the URL to the remote site is served.
If the file,  however, is not found, you are able to serve a mirror to
the site. This example uses two instances of sites to be tested, but
you can expand them to as many as you need.